New Business Frontend Changes
