+++ cd2024bfs2g1-frontend/src/main/ngx/src/app/main/business/business.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { BusinessRoutingModule } from './business-routing.module';
-import { BusinessHomeComponent } from './business-home/business-home.component';
+import { BusinessNewComponent } from './business-new/business-new.component';
 import { BusinessDetailComponent } from './business-detail/business-detail.component';
 import { OFormModule, OntimizeWebModule } from 'ontimize-web-ngx';
 import { OGalleryModule } from 'ontimize-web-ngx-gallery';
@@ -10,7 +10,7 @@ import { OGalleryModule } from 'ontimize-web-ngx-gallery';
 
 @NgModule({
   declarations: [
-    BusinessHomeComponent,
+    BusinessNewComponent,
     BusinessDetailComponent
   ],
   imports: [
